remove unneccesary class
